// apps/mobile-app/src/api/model/dispatchResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Drones API
 * OpenAPI spec version: v1
 */

export interface DispatchResponse {
  id?: string;
  orderId?: string;
  droneId?: string;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  estimatedDeliveryMinutes?: number | null;
  /** @nullable */
  estimatedDeliveryAt?: string | null;
  /** @nullable */
  deliveredAt?: string | null;
  createdAt?: string;
  /** @nullable */
  updatedAt?: string | null;
  droneLatitude?: number;
  droneLongitude?: number;
}